+++ Item.hpp
@@ -14,7 +14,7 @@ private:
 
 public:
     
-    Item(geometrie::Vecteur2<int> p, std::shared_ptr<Image> im, int w, int h, int t)
+    Item(geometrie::Vecteur2<float> p, std::shared_ptr<Image> im, int w, int h, int t)
     : GameObject(p, im, w,h)
     {
         type = t;
